Load main 2015 report card data
Load the data defined in RC15_layout.xlsx and contained in rc15.txt.
This is the school information that is included in every years' report
card, independent of the assessment data, which was shipped separately
in 2015.

Implement a new schema class for the report card data.

Update the value conversion, column type detection, and column naming
logic to accomodate new cases encountered in this data.

Rename the Assessment2015Loader class to DelimitedLoader2015 since it
can be used for both the regular report card data and the assessment
data.

Factor common code between tasks, which was mostly database connection
stuff, into separate functions.